+++ src/main/java/kr/co/teacherforboss/converter/AuthConverter.java
@@ -69,7 +69,7 @@ public class AuthConverter {
 
     public static AuthResponseDTO.ResetPasswordResultDTO toResetPasswordResultDTO(Member member) {
         return AuthResponseDTO.ResetPasswordResultDTO.builder()
-                .memberId(member.getId())
+                .isChanged(true)
                 .changeResult("비밀번호가 재설정 되었습니다.")
                 .build();
     }
