+++ src/main/java/kr/co/teacherforboss/service/authService/AuthCommandServiceImpl.java
@@ -10,6 +10,7 @@ import kr.co.teacherforboss.apiPayload.exception.handler.MemberHandler;
 import kr.co.teacherforboss.config.jwt.PrincipalDetails;
 import kr.co.teacherforboss.config.jwt.TokenManager;
 import kr.co.teacherforboss.converter.AuthConverter;
+import kr.co.teacherforboss.util.PasswordUtil;
 import kr.co.teacherforboss.web.dto.AuthRequestDTO;
 import kr.co.teacherforboss.domain.Member;
 import kr.co.teacherforboss.domain.EmailAuth;
@@ -36,6 +37,7 @@ public class AuthCommandServiceImpl implements AuthCommandService {
     private final MailCommandService mailCommandService;
     private final PasswordEncoder passwordEncoder;
     private final TokenManager tokenManager;
+    private final PasswordUtil passwordUtil;
 
     // 회원 가입
     @Override
@@ -44,30 +46,9 @@ public class AuthCommandServiceImpl implements AuthCommandService {
         if (!request.getPassword().equals(request.getRePassword())) { throw new AuthHandler(ErrorStatus.PASSWORD_NOT_CORRECT);}
 
         Member newMember = AuthConverter.toMember(request);
-        setMemberPassword(newMember, request.getPassword());
+        passwordUtil.setMemberPassword(newMember, request.getPassword());
 
         return memberRepository.save(newMember);
-    }
-
-    // hash 생성
-    private String generatePwHash(String pwRequest, String pwSalt){
-        return passwordEncoder.encode(pwSalt + pwRequest);
-    }
-
-    // salt 생성
-    private String generateSalt() {
-
-        SecureRandom r = new SecureRandom();
-        byte[] salt = new byte[20];
-
-        r.nextBytes(salt);
-
-        StringBuffer sb = new StringBuffer();
-        for(byte b : salt) {
-            sb.append(String.format("%02x", b));
-        };
-
-        return sb.toString();
     }
 
     @Override
@@ -137,14 +118,8 @@ public class AuthCommandServiceImpl implements AuthCommandService {
         if (!request.getPassword().equals(request.getRePassword()))
             throw new AuthHandler(ErrorStatus.PASSWORD_NOT_CORRECT);
 
-        setMemberPassword(member, request.getRePassword());
+        passwordUtil.setMemberPassword(member, request.getRePassword());
 
         return memberRepository.save(member);
-    }
-
-    private void setMemberPassword(Member member, String password) {
-        String pwSalt = generateSalt();
-        String pwHash = generatePwHash(password, pwSalt);
-        member.setPassword(pwSalt, pwHash);
     }
 }
