+++ src/main/java/kr/co/teacherforboss/util/PasswordUtil.java
@@ -1,0 +1,42 @@
+package kr.co.teacherforboss.util;
+
+import kr.co.teacherforboss.domain.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+
+import java.security.SecureRandom;
+
+@Component
+@RequiredArgsConstructor
+public class PasswordUtil {
+
+    private final PasswordEncoder passwordEncoder;
+
+    public void setMemberPassword(Member member, String password) {
+        String pwSalt = generateSalt();
+        String pwHash = generatePwHash(password, pwSalt);
+        member.setPassword(pwSalt, pwHash);
+    }
+
+    // hash 생성
+    private String generatePwHash(String pwRequest, String pwSalt){
+        return passwordEncoder.encode(pwSalt + pwRequest);
+    }
+
+    // salt 생성
+    private String generateSalt() {
+
+        SecureRandom r = new SecureRandom();
+        byte[] salt = new byte[20];
+
+        r.nextBytes(salt);
+
+        StringBuffer sb = new StringBuffer();
+        for(byte b : salt) {
+            sb.append(String.format("%02x", b));
+        };
+
+        return sb.toString();
+    }
+}
