+++ src/main/java/kr/co/teacherforboss/web/dto/AuthResponseDTO.java
@@ -59,7 +59,7 @@ public class AuthResponseDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class ResetPasswordResultDTO {
-        Long memberId;
+        boolean isChanged;
         String changeResult;
     }
 }
